Move select specific upsert implementation into query_builder
