+++ diesel/src/query_builder/upsert/into_conflict_clause.rs
@@ -1,0 +1,125 @@
+use crate::insertable::{BatchInsert, OwnedBatchInsert};
+#[cfg(feature = "sqlite")]
+use crate::query_builder::where_clause::{BoxedWhereClause, WhereClause};
+use crate::query_builder::{
+    AstPass, BoxedSelectStatement, InsertFromSelect, Query, QueryFragment, SelectStatement,
+    ValuesClause,
+};
+
+use crate::result::QueryResult;
+
+pub trait IntoConflictValueClause {
+    type ValueClause;
+
+    fn into_value_clause(self) -> Self::ValueClause;
+}
+
+#[derive(Debug, Clone, Copy)]
+pub struct OnConflictSelectWrapper<S>(S);
+
+impl<Q> Query for OnConflictSelectWrapper<Q>
+where
+    Q: Query,
+{
+    type SqlType = Q::SqlType;
+}
+
+#[cfg(feature = "postgres")]
+impl<S> QueryFragment<crate::pg::Pg> for OnConflictSelectWrapper<S>
+where
+    S: QueryFragment<crate::pg::Pg>,
+{
+    fn walk_ast(&self, out: AstPass<crate::pg::Pg>) -> QueryResult<()> {
+        self.0.walk_ast(out)
+    }
+}
+
+// The corresponding impl for`NoWhereClause` is missing because of
+// https://www.sqlite.org/lang_UPSERT.html (Parsing Ambiguity)
+#[cfg(feature = "sqlite")]
+impl<F, S, D, W, O, L, Of, G, LC> QueryFragment<crate::sqlite::Sqlite>
+    for OnConflictSelectWrapper<SelectStatement<F, S, D, WhereClause<W>, O, L, Of, G, LC>>
+where
+    SelectStatement<F, S, D, WhereClause<W>, O, L, Of, G, LC>: QueryFragment<crate::sqlite::Sqlite>,
+{
+    fn walk_ast(&self, out: AstPass<crate::sqlite::Sqlite>) -> QueryResult<()> {
+        self.0.walk_ast(out)
+    }
+}
+
+#[cfg(feature = "sqlite")]
+impl<'a, ST, QS> QueryFragment<crate::sqlite::Sqlite>
+    for OnConflictSelectWrapper<BoxedSelectStatement<'a, ST, QS, crate::sqlite::Sqlite>>
+where
+    BoxedSelectStatement<'a, ST, QS, crate::sqlite::Sqlite>: QueryFragment<crate::sqlite::Sqlite>,
+    QS: crate::query_source::QuerySource,
+    QS::FromClause: QueryFragment<crate::sqlite::Sqlite>,
+{
+    fn walk_ast(&self, pass: AstPass<crate::sqlite::Sqlite>) -> QueryResult<()> {
+        // https://www.sqlite.org/lang_UPSERT.html (Parsing Ambiguity)
+        self.0.build_query(pass, |where_clause, mut pass| {
+            match where_clause {
+                BoxedWhereClause::None => pass.push_sql("WHERE 1=1"),
+                w => w.walk_ast(pass.reborrow())?,
+            }
+            pass.push_sql(" WHERE 1=1 ");
+            Ok(())
+        })
+    }
+}
+
+impl<Inner, Tab> IntoConflictValueClause for ValuesClause<Inner, Tab> {
+    type ValueClause = Self;
+
+    fn into_value_clause(self) -> Self::ValueClause {
+        self
+    }
+}
+
+impl<'a, Inner, Tab> IntoConflictValueClause for BatchInsert<'a, Inner, Tab> {
+    type ValueClause = Self;
+
+    fn into_value_clause(self) -> Self::ValueClause {
+        self
+    }
+}
+
+impl<Inner, Tab> IntoConflictValueClause for OwnedBatchInsert<Inner, Tab> {
+    type ValueClause = Self;
+
+    fn into_value_clause(self) -> Self::ValueClause {
+        self
+    }
+}
+
+impl<F, S, D, W, O, L, Of, G, LC, Columns> IntoConflictValueClause
+    for InsertFromSelect<SelectStatement<F, S, D, W, O, L, Of, G, LC>, Columns>
+{
+    type ValueClause = InsertFromSelect<
+        OnConflictSelectWrapper<SelectStatement<F, S, D, W, O, L, Of, G, LC>>,
+        Columns,
+    >;
+
+    fn into_value_clause(self) -> Self::ValueClause {
+        let InsertFromSelect { columns, query } = self;
+        InsertFromSelect {
+            query: OnConflictSelectWrapper(query),
+            columns,
+        }
+    }
+}
+
+impl<'a, ST, QS, DB, Columns> IntoConflictValueClause
+    for InsertFromSelect<BoxedSelectStatement<'a, ST, QS, DB>, Columns>
+{
+    type ValueClause =
+        InsertFromSelect<OnConflictSelectWrapper<BoxedSelectStatement<'a, ST, QS, DB>>, Columns>;
+
+    fn into_value_clause(self) -> Self::ValueClause {
+        let InsertFromSelect { columns, query } = self;
+        InsertFromSelect {
+            query: OnConflictSelectWrapper(query),
+            columns,
+        }
+    }
+}
